chore: Rename `api` to `chunithm_net`
