+++ utils/views/compare.py
@@ -8,7 +8,7 @@ from ..components import ScoreCardEmbed
 from .pagination import PaginationView
 
 if TYPE_CHECKING:
-    from api.entities.player_data import PlayerData
+    from chunithm_net.entities.player_data import PlayerData
 
     from ..types.annotated_records import AnnotatedMusicRecord
